Add flag to specify instruction category in elfObjDisasm and singleFileDisasm
