+++ spimdisasm/singleFileDisasm/SingleFileDisasmInternals.py
@@ -36,6 +36,8 @@ def addOptionsToParser(parser: argparse.ArgumentParser) -> argparse.ArgumentPars
 
     parser.add_argument("--split-functions", help="Enables the function and rodata splitter. Expects a path to place the splited functions", metavar="PATH")
 
+    parser.add_argument("--instr-category", help="The instruction category to use when disassembling every passed instruction. Defaults to 'cpu'", choices=["cpu", "rsp", "r5900"])
+
     parser.add_argument("--nuke-pointers", help="Use every technique available to remove pointers", action=common.Utils.BooleanOptionalAction)
     parser.add_argument("--ignore-words", help="A space separated list of hex numbers. Any word differences which starts in any of the provided arguments will be ignored. Max value: FF. Only works when --nuke-pointers is passed", action="extend", nargs="+")
 
@@ -156,6 +158,8 @@ def processArguments(args: argparse.Namespace) -> int:
 
     processedFiles, processedFilesOutputPaths = fec.FrontendUtilities.getSplittedSections(context, splits, array_of_bytes, inputPath, textOutput, dataOutput)
     changeGlobalSegmentRanges(context, processedFiles, len(array_of_bytes), int(args.vram, 16))
+
+    fec.FrontendUtilities.configureProcessedFiles(processedFiles, args.instr_category)
 
     processedFilesCount = 0
     for sect in processedFiles.values():
